+++ source/components/organisms/WatsonAgent.js
@@ -4,8 +4,9 @@ import EventHandler, { EVENT_USER_MESSAGE } from '../../helpers/EventHandler';
 import { sendChatMsg } from '../../services/ChatFormService';
 import ChatBubble from '../atoms/ChatBubble';
 import ButtonStack from '../molecules/ButtonStack';
-import StorageService, { USER_KEY } from "../../services/StorageService";
+import StorageService, { COMPLETED_FORMS_KEY, USER_KEY } from '../../services/StorageService';
 import MarkdownConstructor from "../../helpers/MarkdownConstructor";
+import { Alert, } from "react-native";
 
 let context;
 
@@ -53,6 +54,14 @@ export default class WatsonAgent extends Component {
                     componentProps: {
                         items: [
                             {
+                                value: 'Test service',
+                                action: {
+                                    'type': 'form',
+                                    'value': 2,
+                                    'callback': (params) => this.onFormEnd(params)
+                                }
+                            },
+                            {
                                 value: 'Jag vill boka borgerlig vigsel',
                                 icon: 'wc'
                             },
@@ -64,13 +73,72 @@ export default class WatsonAgent extends Component {
                     }
                 });
             });
-        }
+       }
 
         EventHandler.subscribe(EVENT_USER_MESSAGE, (message) => this.handleHumanChatMessage(message));
     }
 
     componentWillUnmount() {
         EventHandler.unSubscribe(EVENT_USER_MESSAGE);
+    }
+
+    /**
+     * Callback method to run after form is done
+     * @param {object}
+     *
+     * TODO: Move somewhere else?
+     */
+    onFormEnd = async ({ form, answers }) => {
+        const { chat } = this.props;
+
+        const user = await StorageService.getData(USER_KEY);
+
+        const formData = {
+            id: +new Date,
+            userId: user.personalNumber,
+            formId: form.id,
+            created: new Date(),
+            lastUpdated: new Date(),
+            status: 'completed',
+            data: answers,
+        }
+
+        try {
+            await StorageService.putData(COMPLETED_FORMS_KEY, formData);
+        } catch (error) {
+            console.log("Save form error", error);
+        }
+
+        await chat.addMessages([
+            {
+                Component: (props) => <ButtonStack {...props} chat={chat} />,
+                componentProps: {
+                    items: [
+                        {
+                            value: 'Visa mina ärenden',
+                            action: { 'type': 'navigate', 'value': 'UserEvents' },
+                            icon: 'arrow-forward'
+                        },
+                    ]
+                }
+            },
+            {
+                Component: ChatDivider,
+                componentProps: {
+                    info: `Bokning ${form.name.toLowerCase()} avslutad`,
+                }
+            }
+        ]);
+
+        chat.switchAgent(props => <WatsonAgent {...props}
+            initialMessages={['Kan jag hjälpa dig med någon annat?']}
+        />)
+
+        chat.switchInput({
+            autoFocus: false,
+            type: 'text',
+            placeholder: 'Skriv något...'
+        });
     }
 
     /**
@@ -141,7 +209,18 @@ export default class WatsonAgent extends Component {
                     case 'option':
                         const options = current.options.map((option) => {
                             const meta = this.captureMetaData(option.value.input.text);
-                            return { value: option.label, ...meta };
+
+                            const { action } = meta;
+                            // Add callback method to form action
+                            if (action) {
+                                action.callback = action.type === 'form' ? (params) => this.onFormEnd(params) : () => {};
+                            }
+
+                            return {
+                                value: option.label,
+                                ...meta,
+                                action
+                            };
                         });
 
                         const optionType = options[0].type;
